+++ src/hocon/parser/_simple_value.py
@@ -1,7 +1,7 @@
-from typing import Union
+from typing import Union, Optional
 
 from ..constants import ELEMENT_SEPARATORS, SECTION_CLOSING, INLINE_WHITE_CHARS
-from ..exceptions import HOCONUnexpectedSeparatorError, HOCONUnexpectedBracesError
+from ..exceptions import HOCONUnexpectedSeparatorError, HOCONUnexpectedBracesError, HOCONSubstitutionCycleError
 from ._key import parse_keypath
 from ._quoted_string import parse_triple_quoted_string, parse_quoted_string
 from ._unquoted_string import _parse_unquoted_string_value
@@ -9,7 +9,7 @@ from ..strings import UnquotedString, QuotedString
 from ..unresolved import UnresolvedSubstitution
 
 
-def parse_simple_value(data: str, idx: int = 0) -> tuple[Union[UnquotedString, QuotedString, UnresolvedSubstitution], int]:
+def parse_simple_value(data: str, idx: int = 0, current_keypath: Optional[list[str]] = None) -> tuple[Union[UnquotedString, QuotedString, UnresolvedSubstitution], int]:
     char = data[idx]
     if char == ",":
         raise HOCONUnexpectedSeparatorError("Unexpected ',' found.")
@@ -22,7 +22,7 @@ def parse_simple_value(data: str, idx: int = 0) -> tuple[Union[UnquotedString, Q
     elif char in INLINE_WHITE_CHARS:
         return _parse_whitespace_chunk(data, idx)
     elif data[idx:idx + 2] == "${":
-        return _parse_substitution(data, idx + 2)
+        return _parse_substitution(data, idx + 2, current_keypath=current_keypath)
     else:
         return _parse_unquoted_string_value(data, idx)
 
@@ -37,12 +37,15 @@ def _parse_whitespace_chunk(data: str, idx: int) -> tuple[UnquotedString, int]:
         idx += 1
 
 
-def _parse_substitution(data: str, idx: int) -> tuple[UnresolvedSubstitution, int]:
+def _parse_substitution(data: str, idx: int, current_keypath: Optional[list[str]] = None) -> tuple[UnresolvedSubstitution, int]:
     if data[idx] == "?":
         optional = True
         idx += 1
     else:
         optional = False
     keypath = parse_keypath(data, idx, keyend_indicator="}")
-    substitution = UnresolvedSubstitution(keypath.keys, optional)
+    substitution = UnresolvedSubstitution(keypath.keys, optional, location=current_keypath)
+    if current_keypath is not None and ".".join(current_keypath).startswith(".".join(keypath.keys)) and len(current_keypath) > len(keypath.keys):
+        raise HOCONSubstitutionCycleError(
+            f"Substitution {substitution} located at [{'.'.join(current_keypath)}] points to its ancestor node.")
     return substitution, keypath.end_idx